autocast float32

+++ engine.py
@@ -36,7 +36,7 @@ def train_one_epoch(model: torch.nn.Module, criterion: DistillationLoss,
         if mixup_fn is not None:
             samples, targets = mixup_fn(samples, targets)
 
-        with torch.cuda.amp.autocast():
+        with torch.cuda.amp.autocast(dtype=torch.float32):
             outputs = model(samples)
             loss = criterion(samples, outputs, targets)
 
@@ -80,7 +80,7 @@ def evaluate(data_loader, model, device):
         target = target.to(device, non_blocking=True)
 
         # compute output
-        with torch.cuda.amp.autocast():
+        with torch.cuda.amp.autocast(dtype=torch.float32):
             output = model(images)
             loss = criterion(output, target)
