+++ utils.py
@@ -53,7 +53,7 @@ class SmoothedValue(object):
 
     @property
     def avg(self):
-        d = torch.tensor(list(self.deque), dtype=torch.float32)
+        d = torch.tensor(list(self.deque), dtype=torch.float64)
         return d.mean().item()
 
     @property
